Only allow access to DRM_I915_HWS_ADDR ioctl() for Xserver.
CVE-2008-383

The i915_set_status_page() function of the i915 kernel driver calls ioremap
with an address offset that is supplied by userspace via ioctl.
The function zeroes the mapped memory via memset and tells the hardware about
the address. Turns out that access to that ioctl is not restricted to root so
users could probably exploit that to do nasty things. We haven't tried to
write actual exploit code though.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>