UBUNTU: x86: Skip verification by the watchdog for TSC clocksource.
Bug: #319945

Impact: Changes timekeeping on Vmware

This is achieved by resetting the CLOCKSOURCE_MUST_VERIFY flag.

There is a wrap-around problem with the acpi_pm timer.
The acpi_pm counter is just 24bits and this can overflow in ~4 seconds. With
the NO_HZ kernels in virtualized environment, there can be situations when
the guest is descheduled for longer duration, as a result we may miss the wrap
of the acpi counter. When TSC is used as a clocksource and acpi_pm timer is
being used as the watchdog clocksource this error in acpi_pm results in TSC
being marked as unstable, and essentially results in time dropping in chunks
of 4 seconds whenever this wrap is missed. Since the virtualized TSC is
reliable on VMware, we should always use the TSCs clocksource on VMware, so
we skip the verfication at runtime, by checking for the feature bit.

We reset the verify flag for TSC if the X86_FEATURE_TSC_RELIABLE bit is set.

Signed-off-by: Alok N Kataria <akataria@vmware.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>