apparmor: Fix handling of larger number of profiles
Bug: #345144

The profile listing in apparmorfs is broken, so that it will only output
entries that fit on a single page, if there are more profiles than can
be listed in a page worth of memory it bails but when it bails it leaves
some spinlocks held. :(

This breaks, profile listing, replacement and removal if done from the
init scripts.  Manual profile replacement and removal as done by the
utils will still work as long a profile listing has not been done
previously.

Fortunately for this bug only surfaces when a rather large number of
profiles are present, somewhere in the 60-70 profile/hat range.