UBUNTU: SAUCE: do not make sysdev links for processors which are not booted
Bug: #295091

When booting if a CPU cannot be started we will not initialise its sysdev
object.  Later when we initialise ACPI we find both CPUS in the ACPI data
and use that to build sysfs links for each.  When the CPU is not booted
these links have nothing to link to and we panic while building the links.
Do not make or remove the links for CPUs which did not boot.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>