bluetooth hid: enable quirk handling for Apple Wireless Keyboards in 2.6.24
Bug: #227501
BugLink: https://bugs.launchpad.net/ubuntu/+source/linux/+bug/227501

This patch is basically a backport of
commit ee8a1a0a1a5817accd03ced7e7ffde3a4430f485 upstream
and 5a027e910cc26f9b1f8e32daaeb0d552664b2a54 in Intrepid kernel
which was made after the big HID overhaul in 2.6.28.

Kernel 2.6.24 fails to handle quirks for the aluminum Apple Wireless
Keyboard because it is handled as USB device and not as Bluetooth
device. This patch expands 'hidp_blacklist' to make the kernel handle
the keyboard in the same way as the Apple wireless Mighty Mouse (also a
Bluetooth device).

Signed-off-by: Torsten Rausche <torsten@rausche.net>
Cc: Jan Scholz <Scholz@fias.uni-frankfurt.de>
Cc: Jiri Kosina <jkosina@suse.cz>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Bryan Wu <bryan.wu@canonical.com>
Acked-by: Amit Kucheria <amit.kucheria@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>