UBUNTU: SAUCE: init: Add extra mark_rodata_ro quirk for Acer Aspire One
BugLink: https://bugs.launchpad.net/ubuntu/+bug/322867

Fix a very machine and timing dependent problem (as it seems) on the Acer
Aspire One. It looks like calling global_flush_tlb after changing the kernel
data to read only can trigger a kernel panic on this platform. Potentially it
is there for others too, but as it is very timing dependent, other hardware
just might be faster/slower to run into this.
Forcing all other parallel threads into a stop experimentally has solved the
problem and as this path is only run once during boot it will not have much
impact.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>