nfsd4: turn nfsd4_open struct's iattr/verf union into separate fields
CVE-2009-3286

commit 79fb54abd285b442e1f30f851902f3ddf58e7704 upstream (partial
backport)

nfsd4_open has a union that's used to store the verifier for exclusive
creates and the inode attributes for other create modes. The nfsd code
however, will sometimes reference fields in the iattr portion of the
union even for exclusive creates.

Upstream has separated this union into two fields as part of the NFSv4.1
merge.

Signed-off-by: Benny Halevy <bhalevy@panasas.com>
Signed-off-by: J. Bruce Fields <bfields@citi.umich.edu>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>