firewire: ohci: handle receive packets with a data length of zero
CVE-2009-4138

commit 8c0c0cc2d9f4c523fde04bdfe41e4380dec8ee54 upstream (manually
applied)

Queueing to receive an ISO packet with a payload length of zero
silently does nothing in dualbuffer mode, and crashes the kernel in
packet-per-buffer mode.  Return an error in dualbuffer mode, because
the DMA controller won't let us do what we want, and work correctly in
packet-per-buffer mode.

Signed-off-by: Jay Fenlason <fenlason@redhat.com>
Signed-off-by: Stefan Richter <stefanr@s5r6.in-berlin.de>
Cc: stable@kernel.org
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>