UBUNTU: KVM: x86: Completely backport get_cpl functionality
CVE-2010-0306
CVE-2010-0298

For Ubuntu-2.6.24-25.62 (CVE-2009-3290) we partially backported

  commit 2e4d2653497856b102c90153f970c9e344ba96c6 upstream
    KVM: x86: add functions to get the cpl of vcpu

as a prerequisite for

  commit 07708c4af1346ab1521b26a202f438366b7bcffd upstream
    KVM: x86: Disallow hypercalls for guest callers in rings > 0

This seems to have worked on AMD CPUs by luck (probably the call was
never made for those). But now, introduced by

  commit 1871c6020d7308afb99127bba51f04548e7ca84e upstream
    KVM: x86 emulator: fix memory access during x86 emulation

this is called on both CPU types, leading to a blazing crash when tried
on a AMD CPU based system. So this adds the missing part from the
partial backport.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>